+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [1.2.0] - 2021-03-14
 
 ### Added
-- Responsive landing page and favicon.
+- Responsive landing page and favicon. 
 
 ## [1.1.0] - 2021-03-13