+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.3.0] - 2021-03-21
+
+### Added
+- Auth page and integeration with google.
+
 ## [1.2.0] - 2021-03-14
 
 ### Added